Desacoplada/requirements.txt
psycopg2-binary==2.9.11
pydantic==2.11.7